Fixed (#7)
This bug was caused due to inherently different way of handling events in TCP/UDP case and websocket case. In case of TCP/UDP, if initial attempt
to connect to server failed, the client never tried to connect again, whereas websocket handler does try to connect to the server. So, in UDP/TCP
case, the message was popped from the message queue and the send function simply failed and threw a log saying sending message failed as socket
is not connected, whereas in websocket case, it did the same but rather than throwing a error it simply tried to reconnect and sent the new messages.

This can be solved by making the send function blocking, i.e., if websocket not connected then wait forever for it to connect. This commit does exactly
that. Send functions of TCP/UDP are unchanged, but those will be updated in future after going through its problems and rigourous testing.